refactor: replace axios with native fetch api
