+++ src/api/fetch-pokemon-abilities.ts
@@ -1,8 +1,8 @@
 import { Ability } from "@/@types/Ability";
-import { api } from "@/lib/axios";
+import { api } from "@/lib/api";
 
-export async function fetchPokemonAbility(abilityName: string) {
-  const result = await api.get<Ability>(`/ability/${abilityName}`);
-
-  return result.data;
+export async function fetchPokemonAbility(
+  abilityName: string
+): Promise<Ability> {
+  return api(`/ability/${abilityName}`).then((result) => result.json());
 }
